move test files to shared module
